+++ src/Middleware/Handler/HandlerInterface.php
@@ -12,6 +12,7 @@
  */
 
 namespace Gram\Middleware\Handler;
+
 use Gram\Middleware\Classes\ClassInterface;
 
 /**
@@ -19,8 +20,17 @@ use Gram\Middleware\Classes\ClassInterface;
  * @package Gram\Middleware\Handler
  *
  * Ein Interface für Handler die von Middleware ausgeführt werden können
+ *
+ * Wird als Object ausgeführt
+ *
+ * Hat die gleichen Attribute wie Klassen als Strings
  */
 interface HandlerInterface extends ClassInterface
 {
+	/**
+	 * Method die vom Resolver ausgeführt wird
+	 *
+	 * @return mixed
+	 */
 	public function handle();
 }